requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
alembic==1.10.4
